migrate :core:commonui:components to koin annotations
